Moves grid implementation back to a1. Preparing to implement new grid for a2.
