+++ src/app/features/a1/grid/grid.model.ts
@@ -1,9 +1,9 @@
 import * as _ from 'lodash';
 
-import { Surface } from "../surface.model";
-import { SoundName, Note } from "../../sound/sound";
+import { Surface } from "../../../surface/surface.model";
+import { SoundName, Note } from "../../../sound/sound";
 
-import { ticksPerBeat } from "../../features/shared/beat.service";
+import { ticksPerBeat } from "../../shared/beat.service";
 
 let nextId = 0;
