+++ src/app/features/a1/grid/html/beat.component.html
@@ -1,0 +1,8 @@
+<div class="pulse" *ngFor="let pulse of pulseCounts()"
+     [ngClass]="{active: beat.active(thisBeat, pulse, pulses), selected: selected}">
+  <div class="actual note" [ngClass]="noteType()" *ngIf="values[pulse]"></div>
+  <div class="controls">
+    <div class="note" [ngClass]="controlNoteClasses(pulse)"
+         (click)="onToggle(pulse)">{{cursor === pulse ? label : ''}}</div>
+  </div>
+</div>
