+++ src/app/features/a1/grid/html/beat.component.scss
