+++ src/app/features/a1/grid/pixi/pixi-grid.component.scss
