+++ src/app/features/a1/grid/pixi/renderableBar.ts
