+++ src/app/features/a1/grid/pixi/renderableStrip.ts
